pyproject.toml
[project]
name = "dremioframe"
version = "0.5.0"
description = "A dataframe-like library for Dremio Cloud & Dremio Software"
authors = [
    {name = "Alex Merced", email = "alexmerced@alexmerced.com"},
]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyarrow>=14.0.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "polars>=0.20.0",
    "matplotlib",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.8.0",
    "datafusion>=33.0.0",
    "plotly>=5.0.0",
    "pyiceberg>=0.5.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
image_export = ["kaleido"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio",
    "requests-mock"
]
postgres = ["psycopg2-binary"]
mysql = ["mysql-connector-python"]
celery = ["celery", "redis"]
s3 = ["boto3"]
scheduler = ["apscheduler"]
dq = ["pyyaml"]

[project.scripts]
dremio-cli = "dremioframe.cli:app"

[project.urls]
Homepage = "https://github.com/developer-advocacy-dremio/dremio-cloud-dremioframe"
Issues = "https://github.com/developer-advocacy-dremio/dremio-cloud-dremioframe/issues"

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
markers = [
    "cloud: Tests running against Dremio Cloud (requires DREMIO_PAT, DREMIO_PROJECT_ID)",
    "software: Tests running against Dremio Software (requires DREMIO_SOFTWARE_HOST, etc.)",
    "external_backend: Tests running against external backends like Postgres/MySQL (requires DREMIOFRAME_PG_DSN, etc.)"
]
testpaths = ["tests"]
python_files = "test_*.py"
